+++ tests/unit/trace/dsl/test_filter.py
@@ -74,74 +74,74 @@ def test_get_attribute_keys_list(expression: str, expected: Optional[list[str]])
     "expression,expected",
     [
         (
-            "parent_id is not None and 'abc' in name or span_kind == 'LLM' and span_id in ('123',)",  # noqa E501
-            "or_(and_(parent_id != None, TextContains(name, 'abc')), and_(span_kind == 'LLM', span_id.in_(('123',))))"  # noqa E501
+            "parent_id is not None and 'abc' in name or span_kind == 'LLM' and span_id in ('123',)",
+            "or_(and_(parent_id != None, TextContains(name, 'abc')), and_(span_kind == 'LLM', span_id.in_(('123',))))"
             if sys.version_info >= (3, 9)
-            else "or_(and_((parent_id != None), TextContains(name, 'abc')), and_((span_kind == 'LLM'), span_id.in_(('123',))))",  # noqa E501
+            else "or_(and_((parent_id != None), TextContains(name, 'abc')), and_((span_kind == 'LLM'), span_id.in_(('123',))))",
         ),
         (
-            "(parent_id is None or 'abc' not in name) and not (span_kind != 'LLM' or span_id not in ('123',))",  # noqa E501
-            "and_(or_(parent_id == None, not_(TextContains(name, 'abc'))), not_(or_(span_kind != 'LLM', span_id.not_in(('123',)))))"  # noqa E501
+            "(parent_id is None or 'abc' not in name) and not (span_kind != 'LLM' or span_id not in ('123',))",
+            "and_(or_(parent_id == None, not_(TextContains(name, 'abc'))), not_(or_(span_kind != 'LLM', span_id.not_in(('123',)))))"
             if sys.version_info >= (3, 9)
-            else "and_(or_((parent_id == None), not_(TextContains(name, 'abc'))), not_(or_((span_kind != 'LLM'), span_id.not_in(('123',)))))",  # noqa E501
+            else "and_(or_((parent_id == None), not_(TextContains(name, 'abc'))), not_(or_((span_kind != 'LLM'), span_id.not_in(('123',)))))",
         ),
         (
-            "1000 < latency_ms < 2000 or status_code == 'ERROR' or 2000 <= cumulative_llm_token_count_total",  # noqa E501
-            "or_(and_(1000 < latency_ms, latency_ms < 2000), status_code == 'ERROR', 2000 <= cumulative_llm_token_count_total)"  # noqa E501
+            "1000 < latency_ms < 2000 or status_code == 'ERROR' or 2000 <= cumulative_llm_token_count_total",
+            "or_(and_(1000 < latency_ms, latency_ms < 2000), status_code == 'ERROR', 2000 <= cumulative_llm_token_count_total)"
             if sys.version_info >= (3, 9)
-            else "or_(and_((1000 < latency_ms), (latency_ms < 2000)), (status_code == 'ERROR'), (2000 <= cumulative_llm_token_count_total))",  # noqa E501
+            else "or_(and_((1000 < latency_ms), (latency_ms < 2000)), (status_code == 'ERROR'), (2000 <= cumulative_llm_token_count_total))",
         ),
         (
             "llm.token_count.total - llm.token_count.prompt > 1000",
-            "attributes[['llm', 'token_count', 'total']].as_float() - attributes[['llm', 'token_count', 'prompt']].as_float() > 1000"  # noqa E501
+            "attributes[['llm', 'token_count', 'total']].as_float() - attributes[['llm', 'token_count', 'prompt']].as_float() > 1000"
             if sys.version_info >= (3, 9)
-            else "((attributes[['llm', 'token_count', 'total']].as_float() - attributes[['llm', 'token_count', 'prompt']].as_float()) > 1000)",  # noqa E501
+            else "((attributes[['llm', 'token_count', 'total']].as_float() - attributes[['llm', 'token_count', 'prompt']].as_float()) > 1000)",
         ),
         (
             "first.value in (1,) and second.value in ('2',) and '3' in third.value",
-            "and_(attributes[['first', 'value']].as_float().in_((1,)), attributes[['second', 'value']].as_string().in_(('2',)), TextContains(attributes[['third', 'value']].as_string(), '3'))",  # noqa E501
+            "and_(attributes[['first', 'value']].as_float().in_((1,)), attributes[['second', 'value']].as_string().in_(('2',)), TextContains(attributes[['third', 'value']].as_string(), '3'))",
         ),
         (
             "'1.0' < my.value < 2.0",
-            "and_('1.0' < attributes[['my', 'value']].as_string(), attributes[['my', 'value']].as_float() < 2.0)"  # noqa E501
+            "and_('1.0' < attributes[['my', 'value']].as_string(), attributes[['my', 'value']].as_float() < 2.0)"
             if sys.version_info >= (3, 9)
-            else "and_(('1.0' < attributes[['my', 'value']].as_string()), (attributes[['my', 'value']].as_float() < 2.0))",  # noqa E501
+            else "and_(('1.0' < attributes[['my', 'value']].as_string()), (attributes[['my', 'value']].as_float() < 2.0))",
         ),
         (
             "first.value + 1 < second.value",
-            "attributes[['first', 'value']].as_float() + 1 < attributes[['second', 'value']].as_float()"  # noqa E501
+            "attributes[['first', 'value']].as_float() + 1 < attributes[['second', 'value']].as_float()"
             if sys.version_info >= (3, 9)
-            else "((attributes[['first', 'value']].as_float() + 1) < attributes[['second', 'value']].as_float())",  # noqa E501
+            else "((attributes[['first', 'value']].as_float() + 1) < attributes[['second', 'value']].as_float())",
         ),
         (
             "first.value * second.value > third.value",
-            "attributes[['first', 'value']].as_float() * attributes[['second', 'value']].as_float() > attributes[['third', 'value']].as_float()"  # noqa E501
+            "attributes[['first', 'value']].as_float() * attributes[['second', 'value']].as_float() > attributes[['third', 'value']].as_float()"
             if sys.version_info >= (3, 9)
-            else "((attributes[['first', 'value']].as_float() * attributes[['second', 'value']].as_float()) > attributes[['third', 'value']].as_float())",  # noqa E501
+            else "((attributes[['first', 'value']].as_float() * attributes[['second', 'value']].as_float()) > attributes[['third', 'value']].as_float())",
         ),
         (
             "first.value + second.value > third.value",
-            "cast(attributes[['first', 'value']].as_string() + attributes[['second', 'value']].as_string(), String) > attributes[['third', 'value']].as_string()"  # noqa E501
+            "cast(attributes[['first', 'value']].as_string() + attributes[['second', 'value']].as_string(), String) > attributes[['third', 'value']].as_string()"
             if sys.version_info >= (3, 9)
-            else "(cast((attributes[['first', 'value']].as_string() + attributes[['second', 'value']].as_string()), String) > attributes[['third', 'value']].as_string())",  # noqa E501
+            else "(cast((attributes[['first', 'value']].as_string() + attributes[['second', 'value']].as_string()), String) > attributes[['third', 'value']].as_string())",
         ),
         (
             "my.value == '1.0' or float(my.value) < 2.0",
-            "or_(attributes[['my', 'value']].as_string() == '1.0', attributes[['my', 'value']].as_float() < 2.0)"  # noqa E501
+            "or_(attributes[['my', 'value']].as_string() == '1.0', attributes[['my', 'value']].as_float() < 2.0)"
             if sys.version_info >= (3, 9)
-            else "or_((attributes[['my', 'value']].as_string() == '1.0'), (attributes[['my', 'value']].as_float() < 2.0))",  # noqa E501
+            else "or_((attributes[['my', 'value']].as_string() == '1.0'), (attributes[['my', 'value']].as_float() < 2.0))",
         ),
         (
             "not(-metadata['a.b'] + float(metadata[['c.d']]) != metadata[['e.f', 'g.h']])",
-            "not_(-attributes[['metadata', 'a.b']].as_float() + attributes[['metadata', 'c.d']].as_float() != attributes[['metadata', 'e.f', 'g.h']].as_float())"  # noqa E501
+            "not_(-attributes[['metadata', 'a.b']].as_float() + attributes[['metadata', 'c.d']].as_float() != attributes[['metadata', 'e.f', 'g.h']].as_float())"
             if sys.version_info >= (3, 9)
-            else "not_((((- attributes[['metadata', 'a.b']].as_float()) + attributes[['metadata', 'c.d']].as_float()) != attributes[['metadata', 'e.f', 'g.h']].as_float()))",  # noqa E501
+            else "not_((((- attributes[['metadata', 'a.b']].as_float()) + attributes[['metadata', 'c.d']].as_float()) != attributes[['metadata', 'e.f', 'g.h']].as_float()))",
         ),
         (
-            "attributes['attributes'] == attributes[['attributes']] != attributes[['attributes', 'attributes']]",  # noqa E501
-            "and_(attributes[['attributes']].as_string() == attributes[['attributes']].as_string(), attributes[['attributes']].as_string() != attributes[['attributes', 'attributes']].as_string())"  # noqa E501
+            "attributes['attributes'] == attributes[['attributes']] != attributes[['attributes', 'attributes']]",
+            "and_(attributes[['attributes']].as_string() == attributes[['attributes']].as_string(), attributes[['attributes']].as_string() != attributes[['attributes', 'attributes']].as_string())"
             if sys.version_info >= (3, 9)
-            else "and_((attributes[['attributes']].as_string() == attributes[['attributes']].as_string()), (attributes[['attributes']].as_string() != attributes[['attributes', 'attributes']].as_string()))",  # noqa E501
+            else "and_((attributes[['attributes']].as_string() == attributes[['attributes']].as_string()), (attributes[['attributes']].as_string() != attributes[['attributes', 'attributes']].as_string()))",
         ),
     ],
 )
@@ -183,23 +183,23 @@ async def test_filter_translated(
             id="empty-eval-name",
         ),
         pytest.param(
-            """evals['Hallucination'].label == 'correct' or evals['Hallucination'].score < 0.5""",  # noqa E501
-            "span_annotation_0_label_00000000000000000000000000000000 == 'correct' or span_annotation_0_score_00000000000000000000000000000000 < 0.5",  # noqa E501
+            """evals['Hallucination'].label == 'correct' or evals['Hallucination'].score < 0.5""",
+            "span_annotation_0_label_00000000000000000000000000000000 == 'correct' or span_annotation_0_score_00000000000000000000000000000000 < 0.5",
             id="repeated-single-quoted-eval-name",
         ),
         pytest.param(
-            """evals["Hallucination"].label == 'correct' or evals["Hallucination"].score < 0.5""",  # noqa E501
-            "span_annotation_0_label_00000000000000000000000000000000 == 'correct' or span_annotation_0_score_00000000000000000000000000000000 < 0.5",  # noqa E501
+            """evals["Hallucination"].label == 'correct' or evals["Hallucination"].score < 0.5""",
+            "span_annotation_0_label_00000000000000000000000000000000 == 'correct' or span_annotation_0_score_00000000000000000000000000000000 < 0.5",
             id="repeated-double-quoted-eval-name",
         ),
         pytest.param(
-            """evals['Hallucination'].label == 'correct' or evals["Hallucination"].score < 0.5""",  # noqa E501
-            "span_annotation_0_label_00000000000000000000000000000000 == 'correct' or span_annotation_0_score_00000000000000000000000000000000 < 0.5",  # noqa E501
+            """evals['Hallucination'].label == 'correct' or evals["Hallucination"].score < 0.5""",
+            "span_annotation_0_label_00000000000000000000000000000000 == 'correct' or span_annotation_0_score_00000000000000000000000000000000 < 0.5",
             id="repeated-mixed-quoted-eval-name",
         ),
         pytest.param(
-            """evals['Q&A Correctness'].label == 'correct' and evals["Hallucination"].score < 0.5""",  # noqa E501
-            "span_annotation_0_label_00000000000000000000000000000000 == 'correct' and span_annotation_1_score_00000000000000000000000000000000 < 0.5",  # noqa E501
+            """evals['Q&A Correctness'].label == 'correct' and evals["Hallucination"].score < 0.5""",
+            "span_annotation_0_label_00000000000000000000000000000000 == 'correct' and span_annotation_1_score_00000000000000000000000000000000 < 0.5",
             id="distinct-mixed-quoted-eval-names",
         ),
         pytest.param(
@@ -208,23 +208,23 @@ async def test_filter_translated(
             id="missing-right-quotation-mark",
         ),
         pytest.param(
-            """evals["Hallucination"].label == 'correct' orevals["Hallucination"].score < 0.5""",  # noqa E501
-            """span_annotation_0_label_00000000000000000000000000000000 == 'correct' orevals["Hallucination"].score < 0.5""",  # noqa E501
+            """evals["Hallucination"].label == 'correct' orevals["Hallucination"].score < 0.5""",
+            """span_annotation_0_label_00000000000000000000000000000000 == 'correct' orevals["Hallucination"].score < 0.5""",
             id="no-word-boundary-on-the-left",
         ),
         pytest.param(
-            """evals["Hallucination"].scoreq < 0.5""",  # noqa E501
-            """evals["Hallucination"].scoreq < 0.5""",  # noqa E501
+            """evals["Hallucination"].scoreq < 0.5""",
+            """evals["Hallucination"].scoreq < 0.5""",
             id="no-word-boundary-on-the-right",
         ),
         pytest.param(
-            """0.5 <evals["Hallucination"].score""",  # noqa E501
-            """0.5 <span_annotation_0_score_00000000000000000000000000000000""",  # noqa E501
+            """0.5 <evals["Hallucination"].score""",
+            """0.5 <span_annotation_0_score_00000000000000000000000000000000""",
             id="left-word-boundary-without-space",
         ),
         pytest.param(
-            """evals["Hallucination"].score< 0.5""",  # noqa E501
-            """span_annotation_0_score_00000000000000000000000000000000< 0.5""",  # noqa E501
+            """evals["Hallucination"].score< 0.5""",
+            """span_annotation_0_score_00000000000000000000000000000000< 0.5""",
             id="right-word-boundary-without-space",
         ),
         pytest.param(